API/checkers api/requirements.txt
gunicorn==20.1.0
pip~=21.1.2
wheel~=0.36.2
setuptools~=57.0.0
tensorflow-cpu~=2.5.0
numpy~=1.19.5
Flask~=2.0.1
Pillow~=8.2.0